webex-bot/pyproject.toml
[tool.poetry]
name = "webex-bot"
version = "0.1.0"
description = ""
authors = ["fishwongy"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
webex-bot = "^0.4.1"
webexteamssdk = "^1.6.1"
websockets = "^10.2"
emoji = "^2.9.0"
pydantic = "^1.10.6"

google-cloud-secret-manager = "^2.16.3"
google-cloud-logging = "^3.8.0"

requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
